refactor(scripts): extract addTranslationToBasePersonas to be used from the website
